backend/requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
flask-cors==4.0.0
requests==2.31.0
APScheduler==3.10.4
SQLAlchemy==2.0.19
Flask-SQLAlchemy==3.0.5
python-dotenv==1.0.0
urllib3==2.0.7
newsapi-python==0.2.7
